Move interfaces of gadget to more appropriate file.
